Update travis to use official go lint path
Saw that Travis was failing and realized that the problem is with the go lint path. I realized that the lint package was out of date and has changed repositories. This change can be vied in the readme: ead987a65e5c7e053cf9633f9eac1f734f6b4fe3