restapi: retry on 500s, 503s, 504s as well as 502s
This is a sequel to issue #227.

I run a matterbridge bridge and I'm getting 500s from the Discord API
from time to time. We should treat 500, 503, 504 the same as a 502,
instead of just immediately giving up.

Since we're adding entries to the switch statement, also clean up the
naked 429 to be consistent.