Add VerifyUnmarshalInteraction() for performance & convenience
This commit adds VerifyUnmarshalInteraction() which combines
VerifyInteraction() and Interaction.UnmarshalJSON(). This eliminates
the extra data copying that VerifyInteraction() does to put the http
body back into the http request for the caller. Additionally, this
reduces code for callers following the common
verify()/read()/unmarshall() pattern into a single function call.